feat(sql-checker): add assa integration
