+++ modules/fbs-core/api/src/main/resources/application.yml
@@ -67,6 +67,9 @@ services:
   ids:
     salt: ${ID_SALT:feedbacksystem_id_salt}
     length: ${ID_LENGTH:8}
+  assa:
+    url: ${ASSA_URL}
+    token: ${ASSA_TOKEN}
 integrations:
   names: [modelling, feedbackApp, eat, kanban, sciCheck]
   modelling:
