+++ modules/fbs-core/api/src/main/scala/de/thm/ii/fbs/model/checker/assa/Request.scala
@@ -1,0 +1,11 @@
+package de.thm.ii.fbs.model.checker.assa
+
+case class Request(
+  sqlEnvironment: String,
+  dbSchema: String,
+  task: String,
+  solutions: List[String],
+  submissions: List[String],
+  taskId: Option[String],
+  userId: Option[String]
+)
