+++ modules/fbs-core/api/src/main/scala/de/thm/ii/fbs/model/checker/assa/Response.scala
@@ -1,0 +1,3 @@
+package de.thm.ii.fbs.model.checker.assa
+
+case class Response(correct: Boolean, feedback: String)
