+++ modules/fbs-core/web/src/app/dialogs/new-checker-dialog/new-checker-dialog.component.html
@@ -15,6 +15,7 @@
     >
       <mat-option value="sql">SQL Runner</mat-option>
       <mat-option value="sql-checker">SQL Checker</mat-option>
+      <mat-option value="ai-supported-sql-analyser">AI Supported SQL Analyser</mat-option>
       <mat-option value="bash">Bash</mat-option>
       <mat-option value="excel">Excel</mat-option>
     </mat-select>
